+++ messages.go
@@ -17,7 +17,9 @@
 
 package wrp
 
-import "regexp"
+import (
+	"regexp"
+)
 
 //go:generate go install github.com/ugorji/go/codec/codecgen@latest
 //go:generate codecgen -st "json" -o messages_codec.go messages.go