+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add simple message types WRP validators [#85](https://github.com/xmidt-org/wrp-go/pull/85)
+- Add basic WRP spec validators [#84](https://github.com/xmidt-org/wrp-go/pull/84)
+- Introduce WRP Validation Framework [#80](https://github.com/xmidt-org/wrp-go/pull/80)
+- Fix unmarshalling error due to missig metadata fields [#79](https://github.com/xmidt-org/wrp-go/pull/79)
 - Deprecated the concrete message structs, e.g. SimpleEvent
 - Added support for the new qos field.